#263 Invalid interpreter when sharing the .pyproj file in VCS Adds fallback to detection by path names when a virtual environment's base interpreter cannot be found by ID. Moves some virtualenv related functions into DerivedInterpreterFactory.
